Add limit switch code

+++ steering-motor-firmware/Core/Inc/pid.h
@@ -10,7 +10,7 @@
 void resetPID(void);
 void updatePID(void);
 void setPIDGoalD(int16_t distance);
-void setPIDGoalA(int16_t angle);
+void setPIDGoalA(double angle);
 int8_t PIDdone(void); // There is no bool type in C. True/False values are represented as 1 or 0.
 
 #endif /* INC_PID_H_ */
